requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
psycopg2-binary
Flask-WTF
IMDbPY
gunicorn
email-validator